requirements.txt
google-api-python-client==2.29.0
requests==2.26.0
azure-identity==1.7.0
azure-storage-blob==12.8.1
setuptools==58.0.0
Flask==2.0.2
APScheduler==3.9.1
urllib3==1.26.5
tailer==0.4.1
azure-mgmt-resource==21.0.0
azure-mgmt-storage==19.0.0
azure-core==1.19.0
gunicorn==20.1.0
Werkzeug==2.0.2
python-dotenv==0.19.1
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.1
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.3.1
python-dateutil==2.8.2
pytz==2021.3
jsonschema==4.2.1
cryptography==35.0.0